requirements.txt
python-dotenv
google-generativeai==0.5.4